src/galaxymap/requirements.txt
numpy
umap-learn
joblib
scipy
librosa
numba
onnxruntime-gpu
tokenizers
huggingface_hub
nvidia-cuda-runtime-cu12
nvidia-cublas-cu12
nvidia-cufft-cu12
nvidia-curand-cu12
nvidia-cusolver-cu12
nvidia-cusparse-cu12
nvidia-cudnn-cu12
nvidia-cuda-cupti-cu12
vaderSentiment
tinytag
pysqlite3
scikit-learn
threadpoolctl
